Move the api response errors module one level up.
The module `response_errors.py` has been in the folder of arbeitszeit_web/api/presenters. But it's code is used also by controllers and flask_restx views. That's why it has been moved one level up to `arbeitszeit_web/api`.
