+++ arbeitszeit_web/api/controllers/login_member_api_controller.py
@@ -3,7 +3,7 @@ from typing import List
 
 from arbeitszeit.use_cases.log_in_member import LogInMemberUseCase
 from arbeitszeit_web.api.controllers.expected_input import ExpectedInput, InputLocation
-from arbeitszeit_web.api.presenters.response_errors import BadRequest
+from arbeitszeit_web.api.response_errors import BadRequest
 from arbeitszeit_web.request import Request
