+++ arbeitszeit_web/api/presenters/login_member_api_presenter.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 from arbeitszeit.use_cases.log_in_member import LogInMemberUseCase
 from arbeitszeit_web.api.presenters.interfaces import JsonBoolean, JsonObject, JsonValue
-from arbeitszeit_web.api.presenters.response_errors import Unauthorized
+from arbeitszeit_web.api.response_errors import Unauthorized
 from arbeitszeit_web.session import Session
