+++ tests/api/controllers/test_query_plans_api_controller.py
@@ -3,7 +3,7 @@ from arbeitszeit_web.api.controllers.expected_input import InputLocation
 from arbeitszeit_web.api.controllers.query_plans_api_controller import (
     QueryPlansApiController,
 )
-from arbeitszeit_web.api.presenters.response_errors import BadRequest
+from arbeitszeit_web.api.response_errors import BadRequest
 from tests.controllers.base_test_case import BaseTestCase
 from tests.request import FakeRequest
